Added the interfaces to the script language so that polymorphism can be used.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@26 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_scriptengine.cpp
@@ -2947,7 +2947,7 @@ void asCScriptEngine::DeleteScriptFunction(int id)
 	if( scriptFunctions[id] )
 	{
 		delete scriptFunctions[id];
-		if( id == scriptFunctions.GetLength() - 1 )
+		if( id == (int)scriptFunctions.GetLength() - 1 )
 		{
 			scriptFunctions.PopLast();
 		}
